requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
scikit-image>=0.19.0

# Web interface
gradio>=3.20.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Async support
asyncio-mqtt>=0.11.0

# Data validation
pydantic[email]>=2.0.0

# Logging
structlog>=22.0.0

# Additional dependencies for legacy model support
h5py>=3.7.0
cffi>=1.15.0

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Additional API utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Future database support
sqlalchemy>=2.0.0
redis>=5.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
